# web-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trading terminal API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/liquidations": {
      "get": {
        "summary": "Aave V3 liquidations, newest first, cursor-paginated",
        "tags": [
          "liquidations"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
            },
            "in": "query",
            "name": "to",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$"
            },
            "in": "query",
            "name": "asset",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+(\\.\\d+)?$"
            },
            "in": "query",
            "name": "minUsd",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asOf": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "live",
                        "replay",
                        "drill",
                        "offline"
                      ]
                    },
                    "source": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "txHash": {
                            "type": "string"
                          },
                          "logIndex": {
                            "type": "number"
                          },
                          "blockNumber": {
                            "type": "number"
                          },
                          "ts": {
                            "type": "string"
                          },
                          "collateralAsset": {
                            "type": "string"
                          },
                          "collateralSymbol": {
                            "nullable": true,
                            "type": "string"
                          },
                          "debtAsset": {
                            "type": "string"
                          },
                          "debtSymbol": {
                            "nullable": true,
                            "type": "string"
                          },
                          "user": {
                            "type": "string"
                          },
                          "liquidator": {
                            "type": "string"
                          },
                          "debtAmountRaw": {
                            "type": "string"
                          },
                          "collateralAmountRaw": {
                            "type": "string"
                          },
                          "usdValue": {
                            "nullable": true,
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "txUrl": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "txHash",
                          "logIndex",
                          "blockNumber",
                          "ts",
                          "collateralAsset",
                          "collateralSymbol",
                          "debtAsset",
                          "debtSymbol",
                          "user",
                          "liquidator",
                          "debtAmountRaw",
                          "collateralAmountRaw",
                          "usdValue",
                          "status",
                          "txUrl"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "asOf",
                    "mode",
                    "source",
                    "items",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/liquidations/stats": {
      "get": {
        "summary": "Totals, counts, top collateral assets and an hourly series",
        "tags": [
          "liquidations"
        ],
        "parameters": [
          {
            "schema": {
              "default": "24h",
              "type": "string",
              "enum": [
                "1h",
                "24h",
                "7d"
              ]
            },
            "in": "query",
            "name": "window",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asOf": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "live",
                        "replay",
                        "drill",
                        "offline"
                      ]
                    },
                    "source": {
                      "type": "string"
                    },
                    "window": {
                      "type": "string"
                    },
                    "totalUsd": {
                      "type": "string"
                    },
                    "count": {
                      "type": "number"
                    },
                    "topAssets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "asset": {
                            "type": "string"
                          },
                          "symbol": {
                            "nullable": true,
                            "type": "string"
                          },
                          "usd": {
                            "type": "string"
                          },
                          "count": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "asset",
                          "symbol",
                          "usd",
                          "count"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "hourly": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hour": {
                            "type": "string"
                          },
                          "usd": {
                            "type": "string"
                          },
                          "count": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "hour",
                          "usd",
                          "count"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "asOf",
                    "mode",
                    "source",
                    "window",
                    "totalUsd",
                    "count",
                    "topAssets",
                    "hourly"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oracle/{symbol}": {
      "get": {
        "summary": "Chainlink price (what Aave liquidates on) vs the fast market price, and the gap between them",
        "tags": [
          "liquidations"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9]+-[A-Z]+$"
            },
            "in": "path",
            "name": "symbol",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asOf": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "live",
                        "replay",
                        "drill",
                        "offline"
                      ]
                    },
                    "source": {
                      "type": "string"
                    },
                    "oracle": {
                      "type": "object",
                      "properties": {
                        "symbol": {
                          "type": "string"
                        },
                        "chainlinkPrice": {
                          "type": "string"
                        },
                        "fastPrice": {
                          "nullable": true,
                          "type": "string"
                        },
                        "gap": {
                          "nullable": true,
                          "type": "string"
                        },
                        "pressure": {
                          "type": "boolean"
                        },
                        "roundId": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "ageSec": {
                          "type": "number"
                        },
                        "heartbeatSec": {
                          "type": "number"
                        },
                        "deviation": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "symbol",
                        "chainlinkPrice",
                        "fastPrice",
                        "gap",
                        "pressure",
                        "roundId",
                        "updatedAt",
                        "ageSec",
                        "heartbeatSec",
                        "deviation"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "asOf",
                    "mode",
                    "source",
                    "oracle"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}